Fix for the problem that "More" had troubles to display files with more than 32768 lines.
That fixes bug:
https://sourceforge.net/p/aros/bugs/144/

Solved by changing some variable's types from
WORD to LONG.


git-svn-id: https://svn.aros.org/svn/aros/trunk@51553 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/utilities/More/more.c
@@ -137,9 +137,9 @@ static UWORD            tabsize = DEFAULT_TABSIZE;
 
 static WORD             fontwidth, fontheight, borderleft, bordertop;
 static WORD             shinepen, shadowpen, bgpen, textpen;
-static WORD             borderright, borderbottom, visiblex, visibley;
-static WORD             fontbaseline, textstartx, textstarty, textendx, textendy;
-static WORD             textwidth, textheight, viewstartx, viewstarty;
+static LONG             borderright, borderbottom, visiblex, visibley;
+static LONG             fontbaseline, textstartx, textstarty, textendx, textendy;
+static LONG             textwidth, textheight, viewstartx, viewstarty;
 static WORD             winwidth, winheight;
 
 static ULONG            winmask, msgmask;
